feat #8 Adding development mode
This commit add the possibility to run the snippets server locally targeting your local version of documentation files. It gives you the ability to work locally on samples and snippets before push them to GH.
It also disable the cache when you are in dev mode and load framework and skin from the cdn.